Improve resources code layout
