+++ superdesk/core/resources/types.py
@@ -1,0 +1,5 @@
+from typing import TypeVar
+from .model import ResourceModel
+
+
+ResourceModelType = TypeVar("ResourceModelType", bound=ResourceModel)
